+++ arch/arm/mach-k3/common.h
@@ -48,8 +48,8 @@ void do_board_detect(void);
 void ti_secure_image_check_binary(void **p_image, size_t *p_size);
 void wkup_ctrl_remove_can_io_isolation_if_set(void);
 bool wkup_ctrl_is_lpm_exit(void);
-int wkup_r5f_am62_lpm_meta_data_addr(u64 *meta_data_addr);
-void lpm_resume_from_ddr(u64 meta_data_addr);
+int wkup_r5f_am62_lpm_meta_data_addr(u32 *meta_data_addr);
+void lpm_resume_from_ddr(u32 meta_data_addr);
 int shutdown_mcu_r5_core1(void);
 
 #if (IS_ENABLED(CONFIG_K3_QOS))
